Fixed all errors in CSVExport (now compiles), wrote a test for CSVExport, added file comparison utility

+++ SimulationLib/CSVExport-inl.h
@@ -104,7 +104,7 @@ namespace SimulationLib {
 
         // Perform write
         stringSize = buf.size();
-        if (fwrite(buf.c_str(), sizeof(char), stringSize, fs) < stringSize) {
+        if (fwrite(buf.c_str(), sizeof(char), stringSize, fs) < (size_t)stringSize) {
             printf("Error: std::fwrite returned an error\n");
             fclose(fs);
             return;
